Style correction and workflow simplification
Kernel.cpp has a simpler print function
Removed constructors pointer until needed
Improved code readability (following my style now)
Improved makefile
Redacted readme

+++ src/types.h
@@ -8,10 +8,10 @@
 	typedef short int16_t;
 	typedef unsigned short uint16_t;
 	
-	typedef short int32_t;
-	typedef unsigned short uint32_t;
+	typedef int int32_t;
+	typedef unsigned int uint32_t;
 
-	typedef short int64_t;
-	typedef unsigned short uint64_t;
+	typedef long long int64_t;
+	typedef unsigned long long uint64_t;
 
 #endif
